Funcionalidad login y zona privada personalizada al usuario logged

+++ MenuApp.cs
@@ -4,7 +4,6 @@ using Services;
 
 class MenuApp
 {
-
     public MenuApp()
     {
         
@@ -20,11 +19,28 @@ class MenuApp
             Console.WriteLine("\n--- MENÚ PRINCIPAL ---");
             Console.WriteLine("1. Géneros");
             Console.WriteLine("2. Comics");
-            Console.WriteLine("3. Usuarios");
-            Console.WriteLine("4. Salir");
+            
+            if (UserService.currentUser != null && UserService.currentUser.IsAdmin)
+            {
+                Console.WriteLine("3. Usuarios");
+            }
+
+            if (UserService.currentUser != null)
+            {
+                Console.WriteLine("4. Zona privada");
+                Console.WriteLine("5. Cerrar sesión");
+                Console.WriteLine("6. Salir");
+            }
+            else
+            {
+                Console.WriteLine("4. Iniciar sesión");
+                Console.WriteLine("5. Registrarse");
+                Console.WriteLine("6. Salir");
+            }
+
             Console.WriteLine("Selecciona una opción:");
 
-            if (!int.TryParse(Console.ReadLine(), out option) || option < 1 || option > 4)
+            if (!int.TryParse(Console.ReadLine(), out option) || option < 1 || option > 6)
             {
                 Console.WriteLine("Error: Por favor selecciona una opción válida (1-6).");
                 continue;
@@ -42,6 +58,18 @@ class MenuApp
                     ShowUserMenu();
                     break;
                 case 4:
+                    if (UserService.currentUser != null)
+                        ShowPrivateMenu();
+                    else
+                        UserService.Login();
+                    break;
+                case 5:
+                    if (UserService.currentUser != null)
+                        UserService.Logout();
+                    else
+                        UserService.AddUser();
+                    break;
+                case 6:
                     Console.WriteLine("¡Hasta pronto!");
                     break;
                 default:
@@ -50,7 +78,7 @@ class MenuApp
             }
         }
 
-        while(option != 4);
+        while(option != 6);
     }
 
 
@@ -79,7 +107,10 @@ class MenuApp
             switch (option)
             {
                 case 1:
-                    GenreService.AddGenre();
+                    if (UserService.currentUser != null)
+                        GenreService.AddGenre();
+                    else
+                        Console.WriteLine("Debes iniciar sesión para realizar esta acción.");
                     break;
                 case 2:
                     GenreService.ShowAllGenres();
@@ -88,7 +119,10 @@ class MenuApp
                     GenreService.SearchGenre();
                     break;
                 case 4:
-                    GenreService.DeleteGenre();
+                    if (UserService.currentUser != null)
+                        GenreService.DeleteGenre();
+                    else
+                        Console.WriteLine("Debes iniciar sesión para realizar esta acción.");
                     break;
                 case 5:
                     Console.WriteLine("Volviendo..");
@@ -111,17 +145,15 @@ class MenuApp
 
         do
         {
-            Console.WriteLine("\n--- MENÚ CÓMIC ---");
+            Console.WriteLine("\n--- MENÚ COMICS ---");
             Console.WriteLine("1. Añadir cómic");
             Console.WriteLine("2. Listar cómics");
             Console.WriteLine("3. Buscar cómic");
             Console.WriteLine("4. Eliminar cómic");
-            Console.WriteLine("5. Añadir cómic a lista personal");
-            Console.WriteLine("6. Otro");
-            Console.WriteLine("7. Volver al menú principal");
+            Console.WriteLine("5. Volver al menú principal");
             Console.WriteLine("Selecciona una opción:");
 
-            if (!int.TryParse((string)Console.ReadLine(), out option) || option < 1 || option > 7)
+            if (!int.TryParse((string)Console.ReadLine(), out option) || option < 1 || option > 5)
             {
                 Console.WriteLine("Error: Por favor selecciona una opción válida (1-7).");
                     continue;
@@ -130,7 +162,10 @@ class MenuApp
             switch (option)
             {
                 case 1:
-                    ComicService.AddComic();
+                    if (UserService.currentUser != null)
+                        ComicService.AddComic();
+                    else
+                        Console.WriteLine("Debes iniciar sesión para realizar esta acción.");
                     break;
                 case 2:
                     ComicService.ShowAllComics();
@@ -139,15 +174,12 @@ class MenuApp
                     ComicService.SearchComic();
                     break;
                 case 4:
-                    ComicService.DeleteComic();
+                    if (UserService.currentUser != null)
+                        ComicService.DeleteComic();
+                    else
+                        Console.WriteLine("Debes iniciar sesión para realizar esta acción.");
                     break;
                 case 5:
-                    UserService.AddComicToUserList();
-                    break;
-                case 6:
-                    //TODO
-                    break;
-                case 7:
                     Console.WriteLine("Volviendo..");
                     break;
                 default:
@@ -156,7 +188,7 @@ class MenuApp
             }
 
         }
-        while (option != 7);
+        while (option != 5);
 
     }
 
@@ -208,5 +240,48 @@ class MenuApp
         }
         while (option != 5);
 
+    }
+
+
+
+    private void ShowPrivateMenu()
+    {
+        int option = 0;
+
+        do
+        {
+            Console.WriteLine("\n--- ZONA PRIVADA ---");
+            Console.WriteLine("1. Añadir cómic a mi lista personal");
+            Console.WriteLine("2. Eliminar cómic a mi lista personal");
+            Console.WriteLine("3. Ver mi lista personal de cómics");
+            Console.WriteLine("4. Volver al menú principal");
+
+            Console.WriteLine("Selecciona una opción:");
+
+            if (!int.TryParse(Console.ReadLine(), out option) || option < 1 || option > 4)
+            {
+                Console.WriteLine("Error: Por favor selecciona una opción válida (1-4).");
+                continue;
+            }
+
+            switch (option)
+            {
+                case 1:
+                    UserService.ManageComicsInUserList(UserService.currentUser, true);
+                    break;
+                case 2:
+                    UserService.ManageComicsInUserList(UserService.currentUser, false);
+                    break;
+                case 3:
+                    UserService.ShowUserComics(UserService.currentUser);
+                    break;
+                case 4:
+                    Console.WriteLine("Volviendo...");
+                    break;
+                default:
+                    Console.WriteLine("La opción no es correcta");
+                    break;
+            }
+        } while (option != 4);
     }
 }
